INT-83 Unify types into one file
* Use `index.d.ts` instead of multiple files
